requirements.txt
django==4.2.7
python-decouple==3.8
pre-commit==3.3.3
python-magic==0.4.27
djangorestframework==3.14.0
